// apps/admin/src/types/page-editor.ts
export interface Component {
  id: string;
  type: string;
  label: string;
  props?: Record<string, any>;
}

export interface LayoutComponent extends Component {
  id: string;
  width?: number;
  height?: number;
}